# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Market Backend API
  version: 0.1.0
  description: 'Unified HTTP API contract for Market Backend. Operations marked with `x-status: implemented` are
    exposed by the current FastAPI application; operations marked with `x-status: planned` are contract-first proposals
    and are not available yet.'
paths:
...
  /api/v1/admin/products:
    get:
      tags:
      - admin
      summary: List Products
      operationId: list_products_api_v1_admin_products_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
                title: Response List Products Api V1 Admin Products Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
    post:
      tags:
      - admin
      summary: Create Product
      operationId: create_product_api_v1_admin_products_post
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/admin/products/{product_id}:
    get:
      tags:
      - admin
      summary: Get Product
      operationId: get_product_api_v1_admin_products__product_id__get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
    patch:
      tags:
      - admin
      summary: Update Product
      operationId: update_product_api_v1_admin_products__product_id__patch
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
    delete:
      tags:
      - admin
      summary: Delete Product
      operationId: delete_product_api_v1_admin_products__product_id__delete
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
  /api/v1/products:
    get:
      tags:
      - products
      summary: List Products
      description: Return a bounded page of products without requiring authentication.
      operationId: list_products_api_v1_products_get
      parameters:
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PublicProductResponse'
                title: Response List Products Api V1 Products Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/products/{product_id}:
    get:
      tags:
      - products
      summary: Get Product
      description: Return one public product card or the common domain-level 404 response.
      operationId: get_product_api_v1_products__product_id__get
      parameters:
      - name: product_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Product Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicProductResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
  /api/v1/orders/{order_id}/payments/tbank:
    post:
      tags:
      - payments
      summary: Create Tbank Payment
      operationId: create_tbank_payment_api_v1_orders__order_id__payments_tbank_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTBankPaymentRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/orders/{order_id}/payment:
    get:
      tags:
      - payments
      summary: Get Order Payment
      operationId: get_order_payment_api_v1_orders__order_id__payment_get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
  /api/v1/orders/{order_id}/payment/cancel:
    post:
      tags:
      - payments
      summary: Cancel Order Payment
      operationId: cancel_order_payment_api_v1_orders__order_id__payment_cancel_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: Authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      x-status: implemented
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProductResponse:
      properties:
        uuid:
          type: string
          format: uuid
          title: Uuid
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        images:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Images
        header_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Header Image
        price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Price
        ozon_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Ozon Price
        wb_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Wb Price
        tags:
          items:
            $ref: '#/components/schemas/TagSummary'
          type: array
          title: Tags
        created_by:
          type: string
          format: uuid
          title: Created By
        updated_by:
          type: string
          format: uuid
          title: Updated By
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - uuid
      - title
      - description
      - images
      - header_image
      - price
      - ozon_price
      - wb_price
      - tags
      - created_by
      - updated_by
      - created_at
      - updated_at
      title: ProductResponse
    PublicProductResponse:
      properties:
        uuid:
          type: string
          format: uuid
          title: Uuid
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        images:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Images
        header_image:
          anyOf:
          - type: string
          - type: 'null'
          title: Header Image
        price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Price
        ozon_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Ozon Price
        wb_price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Wb Price
        tags:
          items:
            $ref: '#/components/schemas/TagSummary'
          type: array
          title: Tags
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - uuid
      - title
      - description
      - images
      - header_image
      - price
      - ozon_price
      - wb_price
      - tags
      - created_at
      - updated_at
      title: PublicProductResponse
      description: Storefront representation deliberately excludes administrator identifiers.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    CreateTBankPaymentRequest:
      properties:
        payment_option_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Payment Option Id
        idempotency_key:
          type: string
          maxLength: 64
          minLength: 1
          title: Idempotency Key
      additionalProperties: false
      type: object
      required:
      - payment_option_id
      - idempotency_key
      title: CreateTBankPaymentRequest
...
    PaymentConfirmationResponse:
      properties:
        type:
          type: string
          enum:
          - redirect
          - qr
          title: Type
        url:
          type: string
          title: Url
      additionalProperties: false
      type: object
      required:
      - type
      - url
      title: PaymentConfirmationResponse
    PaymentResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        order_id:
          type: string
          format: uuid
          title: Order Id
        provider:
          type: string
          title: Provider
        payment_option_id:
          type: string
          title: Payment Option Id
        status:
          type: string
          title: Status
        amount:
          type: string
          title: Amount
        currency:
          type: string
          title: Currency
        confirmation:
          anyOf:
          - $ref: '#/components/schemas/PaymentConfirmationResponse'
          - type: 'null'
        failure_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Failure Message
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      additionalProperties: false
      type: object
      required:
      - id
      - order_id
      - provider
      - payment_option_id
      - status
      - amount
      - currency
      - confirmation
      - failure_message
      - created_at
      - updated_at
      title: PaymentResponse
...
    TagSummary:
      properties:
        uuid:
          type: string
          format: uuid
          title: Uuid
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
      additionalProperties: false
      type: object
      required:
      - uuid
      - name
      - slug
      title: TagSummary
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
servers:
- url: /
  description: Current origin
...